+++ src/main/webapp/resources/views/details.jsp
@@ -29,36 +29,25 @@
     </div>
 </header>
 
-
 <div class="details">
-    <div class="details-apprenticeInformation">
-
-    </div>
-    <div class="details-apprenticeInformation">
-
-    </div>
-    <div class="details-apprenticeInformation">
-
-    </div>
-
     <div class="input-row">
         <div class="half-width flex">
             <span class="bold">Lastname</span>
-            <span>${mainUser.lastname}</span>
+            <span>${user.lastname}</span>
         </div>
         <div class="half-width flex">
             <span class="bold">Forename</span>
-            <span>${mainUser.forename}</span>
+            <span>${user.forename}</span>
         </div>
     </div>
     <div class="input-row">
         <div class="half-width flex">
             <span class="bold">Email</span>
-            <span>${mainUser.email}</span>
+            <span>${user.email}</span>
         </div>
         <div class="half-width flex">
             <span class="bold">Phone Number</span>
-            <span>${mainUser.phone}</span>
+            <span>${user.phone}</span>
         </div>
     </div>
     <hr />
@@ -99,24 +88,36 @@
             </div>
             <div class="half-width flex">
                 <span class="bold">Tutor</span>
-                <span><a href="details?userId=${apprenticeTutor.userId}">${apprenticeTutor.forename} ${apprenticeTutor.lastname}</a></span>
+                <span><a href="detail?userId=${apprenticeTutor.userId}">${apprenticeTutor.forename} ${apprenticeTutor.lastname}</a></span>
             </div>
         </div>
         <button class="submit" onclick="location.href='${pageContext.request.contextPath}/update-apprentice?idApprentice=${associatedApprentice.idApprentice}';">Update Apprentice's Information</button>
     </c:if>
 
     <c:if test="${associatedTutor != null}">
-        <h3 class="popup-title">Tutor</h3>
-        <div class="input-row">
-            <div class="half-width flex">
-                <span class="bold">Address</span>
-                <span>4 rue des Chasseurs</span>
-            </div>
-            <div class="half-width flex">
-                <span class="bold">Infos</span>
-                <span>A la cambrousse</span>
-            </div>
-        </div>
+        <c:if test="${not empty tutorApprentices}">
+            <table>
+                <thead>
+                <tr>
+                    <th>Name</th>
+                    <th>Forename</th>
+                    <th>Role</th>
+                    <th></th>
+                </tr>
+                </thead>
+                <c:forEach var="apprentice" items="${tutorApprentices}">
+                    <tr>
+                        <td>${apprentice.lastname}</td>
+                        <td>${apprentice.forename}</td>
+                        <td>${apprentice.userType}</td>
+                        <td><button class="submit" onclick="location.href='${pageContext.request.contextPath}/detail?userId=${apprentice.userId}';">Details</button></td>
+                    </tr>
+                </c:forEach>
+            </table>
+        </c:if>
+        <c:if test="${empty tutorApprentices}">
+
+        </c:if>
     </c:if>
 
     <c:if test="${associatedApprentice == null && associatedTutor == null}">
@@ -125,12 +126,6 @@
 
 
 </div>
-
-
-
-
-
-
 
 </body>
 </html>
